+++ src/util/Crypto.hpp
@@ -8,6 +8,8 @@ namespace chatterino {
 
 QString encryptMessage(QStringView message, QStringView encryptionPassword);
 
-bool checkAndDecryptMessage(QString &message, QStringView encryptionPassword);
+bool decryptMessage(QString &message, QStringView encryptionPassword);
+
+bool isMaybeEncrypted(QStringView message);
 
 }  // namespace chatterino